More fixes to large file size in MPI -- still not working though

+++ mympi.global.funcdeclare.h
@@ -75,7 +75,7 @@ extern void mpiisum0(int*sumptr, int recvid);
 extern void mpildsum0(long int*sumptr, int recvid);
 extern void mpifmin(FTYPE*minptr);
 
-extern int set_sizeofmemory(int numbuff, int sizeofdatatype, int numcolumns, int *sizeofmemory);
+extern int set_sizeofmemory(int numbuff, int sizeofdatatype, int numcolumns, long long int *sizeofmemory);
 extern int set_maxnumsize(int numcolumns, long long int *maxnumsize);
 extern int set_numbuffers(int numcolumns, int *numbuffers);
 extern int gcountmod(int numcolumns);
